Debugging && add bonus rules to makefile

+++ sources/conversions_utils_bonus.c
@@ -6,7 +6,7 @@
 /*   By: mperrine <mperrine@student.42angouleme.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/07 13:23:28 by mperrine          #+#    #+#             */
-/*   Updated: 2026/03/07 14:24:15 by mperrine         ###   ########.fr       */
+/*   Updated: 2026/03/07 20:55:17 by mperrine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	apply_width(char **src, char fill, t_ft_printf *data)
 {
-	size_t	i;
-	size_t	srclen;
+	int		i;
+	int		srclen;
 	char	*dest;
 
 	srclen = ft_strlen(*src);
@@ -43,9 +43,9 @@ int	apply_width(char **src, char fill, t_ft_printf *data)
 int	apply_precision_nb(char **src, t_ft_printf *data)
 {
 	char	*dest;
-	size_t	size;
-	size_t	srclen;
-	size_t	i;
+	int		size;
+	int		srclen;
+	int		i;
 
 	srclen = ft_strlen(*src);
 	if (data->flags.prec == -1 || data->flags.prec <= srclen)
@@ -73,7 +73,7 @@ int	apply_precision_str(char **src, t_ft_printf *data)
 {
 	char	*dest;
 
-	if (data->flags.prec == -1 || data->flags.prec >= ft_strlen(*src))
+	if (data->flags.prec == -1 || data->flags.prec >= (int)ft_strlen(*src))
 		return (0);
 	dest = calloc(data->flags.prec + 1, sizeof(char));
 	if (!dest)
