+++ sources/maths_utils_bonus.c
@@ -6,7 +6,7 @@
 /*   By: mperrine <mperrine@student.42angouleme.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/29 17:33:24 by mperrine          #+#    #+#             */
-/*   Updated: 2026/03/07 13:19:09 by mperrine         ###   ########.fr       */
+/*   Updated: 2026/03/07 23:27:52 by mperrine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,16 +32,16 @@ void	set_nb_prefix(char *s, long *nb, t_ft_printf *data)
 	}
 }
 
-size_t	get_hex_size(size_t nb)
+int	get_hex_size(size_t nb)
 {
 	if (nb < 16)
 		return (1);
 	return (1 + get_hex_size(nb / 16));
 }
 
-size_t	get_dec_size(long nb)
+int	get_dec_size(long nb)
 {
-	size_t	sign;
+	int	sign;
 
 	sign = 0;
 	if (nb < 0)
